requirements.txt
django
django-rest-framework==0.1.0
djangorestframework==3.7.7